backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
pypdf
nltk
python-dotenv
